+++ packages/wrapped-keys-lit-actions/src/lib/self-executing-actions/ethereum/signMessageWithEncryptedEthereumKey.js
@@ -1,47 +1,11 @@
-const {
-  getDecryptedKeyToSingleNode,
-} = require('../../internal/common/getDecryptedKeyToSingleNode');
-const {
-  signMessageEthereumKey,
-} = require('../../internal/ethereum/signMessage');
-const { removeSaltFromDecryptedKey } = require('../../utils');
+/* global accessControlConditions, ciphertext, dataToEncryptHash, messageToSign */
 
-/* global accessControlConditions, ciphertext, dataToEncryptHash, messageToSign, Lit */
+import { signMessageWithEncryptedEthereumKey } from '../../raw-action-functions/ethereum/signMessageWithEncryptedEthereumKey';
 
-/**
- * Signs a message with the Ethers wallet which is also decrypted inside the Lit Action.
- *
- * @jsParam pkpAddress - The Eth address of the PKP which is associated with the Wrapped Key
- * @jsParam ciphertext - For the encrypted Wrapped Key
- * @jsParam dataToEncryptHash - For the encrypted Wrapped Key
- * @jsParam messageToSign - The unsigned message to be signed by the Wrapped Key
- * @jsParam accessControlConditions - The access control condition that allows only the pkpAddress to decrypt the Wrapped Key
- *
- * @returns { Promise<string> } - Returns a message signed by the Ethers Wrapped key. Or returns errors if any.
- */
-
-(async () => {
-  try {
-    const decryptedPrivateKey = await getDecryptedKeyToSingleNode({
-      accessControlConditions,
-      ciphertext,
-      dataToEncryptHash,
-    });
-
-    if (!decryptedPrivateKey) {
-      // Silently exit on nodes which didn't run the `decryptToSingleNode` code
-      return;
-    }
-
-    const privateKey = removeSaltFromDecryptedKey(decryptedPrivateKey);
-
-    const signature = await signMessageEthereumKey({
-      privateKey,
-      messageToSign,
-    });
-
-    Lit.Actions.setResponse({ response: signature });
-  } catch (err) {
-    Lit.Actions.setResponse({ response: `Error: ${err.message}` });
-  }
-})();
+(async () =>
+  signMessageWithEncryptedEthereumKey({
+    accessControlConditions,
+    ciphertext,
+    dataToEncryptHash,
+    messageToSign,
+  }))();
